Fixed some leakage of database connections.
Reworked handling of bibtex's entries that are not in English. It is still compatible with the old format, but we now consider, by default, that title, abstract and keyword are in English (instead of the original language). Original language title, abstract and keyword are now stored in title-orig, abstract-orig and keywords-orig.